# packages/contracts/openapi/alo-noon.v1.yaml
openapi: 3.1.0
info:
  title: Alo Noon API
  version: 0.11.0
  description: >-
    Phase 2E contract. Operational endpoints, tenant-aware discovery, identity,
    customer-owned addresses, authoritative delivery quotes, and atomic
    capacity-backed Quote-to-Order conversion, and the internal payment and
    double-entry ledger foundation are implemented. Payment execution remains
    internal; the versioned tenant chart of accounts and governed account state
    foundation are implemented. The tenant-scoped, SPI-versioned payment-provider abstraction,
    opaque credential-reference, payment-attempt, and callback replay foundation
    are implemented without a real gateway adapter. A customer may open the payment for an
    order they own and read it back, and nothing more: no public endpoint lets a client name
    a payment amount or assert a payment state, and no ledger-governance write endpoint is
    exposed at all. The authenticated initialization orchestrator is implemented as an
    injectable, fail-closed boundary. Gateway callbacks are recorded and then settled by
    server-to-server verification: a payment is captured, posted to the double-entry ledger,
    and marked paid on its order only when the gateway confirms it for the exact expected
    amount. No settlement surface is exposed publicly — the customer's return carries an
    opaque reference and no verdict. Authentication delivery
...
servers:
  - url: http://localhost:3001
paths:
...
  /api/v1/admin/access/roles:
    get:
      operationId: listAdminAccessRoles
      summary: The role catalogue, marked with what this caller may grant.
      description: >-
        `grantable` is false when the role carries a permission the caller does not itself
        hold. Roles are still listed, so an operator can see what they would need rather
        than wondering why a role is missing.
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      responses:
        '200':
          description: Every admin role.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminRoleListEnvelope' }
        '401': { description: A valid staff session is required. }
        '403': { description: The account may not manage staff access. }
  /api/v1/admin/access/staff:
    get:
      operationId: listAdminStaff
      summary: Accounts in this tenant that hold, or have held, an admin role.
      description: >-
        Deliberately not every tenant member. The tenant's customers are identity accounts
        too, and listing them here would turn an access screen into a customer directory.
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      responses:
        '200':
          description: Staff accounts with their live grants.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StaffMemberListEnvelope' }
        '401': { description: A valid staff session is required. }
        '403': { description: The account may not manage staff access. }
        '503': { description: Access management is temporarily unavailable. }
  /api/v1/admin/access/grants:
    post:
      operationId: grantAdminRole
      summary: Grant a role to a mobile number.
      description: >-
        Refused when the role carries any permission the caller does not hold, measured
        against live grant rows rather than the caller's session. The target must have
        signed in once with that number — signing in is what creates the account and what
        proves whoever holds the number asked to be here. Re-granting a role someone
        already holds is a no-op, not an error.
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/GrantRoleCommand' }
      responses:
        '201':
          description: The staff member, with their grants after the change.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StaffMemberEnvelope' }
        '400': { description: The command is malformed, or names an unknown role. }
        '401': { description: A valid staff session is required. }
        '403': { description: The account may not manage access, or the role exceeds its own. }
        '404': { description: No account for that number. }
        '409': { description: That account is not active, or not a member of this tenant. }
        '503': { description: Access management is temporarily unavailable. }
  /api/v1/admin/access/revocations:
    post:
      operationId: revokeAdminRole
      summary: Revoke a role grant.
      description: >-
        Measured by the same yardstick as granting: revoking a role the caller could not
        have granted is refused, because otherwise a narrow operator could strip every
        wider one and be left alone in the tenant. An account cannot revoke its own grant —
        recovering from that needs shell access to the database server.
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RevokeRoleCommand' }
      responses:
        '200':
          description: The staff member, with their grants after the change.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StaffMemberEnvelope' }
        '400': { description: The command is malformed. }
        '401': { description: A valid staff session is required. }
        '403': { description: The account may not manage access, or the role exceeds its own. }
        '404': { description: No such grant in this tenant. }
        '409': { description: An account cannot revoke its own access. }
        '503': { description: Access management is temporarily unavailable. }
  /api/v1/admin/access/branches:
    get:
      operationId: listGrantableBranches
      summary: The branches a branch-scoped role can be granted against.
      description: >-
        Answers to admin.access.manage rather than to the catalogue's own permission.
        ACCESS_ADMIN — the role whose entire job is issuing grants — does not hold
        admin.catalog.manage, so borrowing the catalogue's branch listing would leave
        exactly that role unable to fill in the one field a branch grant needs. Three
        fields, because everything else about a branch is the catalogue's business and
        none of it helps somebody pick one from a list.
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      responses:
        '200':
          description: Branches in this tenant, by name.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GrantableBranchListEnvelope' }
        '401': { description: A valid staff session is required. }
        '403': { description: The account may not manage access. }
        '503': { description: Access management is temporarily unavailable. }
  /api/v1/admin/settlement/outstanding:
    get:
      operationId: listOutstandingPartnerBalances
      summary: What each partner has earned and not been paid.
      description: >-
        Summed in the database rather than read and added here: a month of one city's
        orders is a great many rows to move across a socket to produce one number per
        partner, and the sum has to be exact. `oldestAt` is the figure a total alone
        hides — two partners owed the same amount are in very different positions if one
        has been waiting since last month.
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      responses:
        '200':
          description: Every partner with something owing, largest first.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PartnerBalanceListEnvelope' }
        '401': { description: A valid staff session is required. }
        '403': { description: The account may not settle partner payouts. }
        '503': { description: The payout desk is temporarily unavailable. }
  /api/v1/admin/settlement/payouts:
    get:
      operationId: listPartnerPayouts
      summary: Payout runs, newest first.
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters:
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
      responses:
        '200':
          description: The payout runs this tenant has prepared.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PartnerPayoutListEnvelope' }
        '400': { description: That payout listing is not a valid request. }
        '401': { description: A valid staff session is required. }
        '403': { description: The account may not settle partner payouts. }
        '503': { description: The payout desk is temporarily unavailable. }
    post:
      operationId: preparePartnerPayout
      summary: Prepare a payout covering everything a partner is owed.
      description: >-
        The amount is never in the request: it is whatever the partner has earned and not
        been paid at the instant the transaction reads it, under a row lock, so two runs
        prepared at the same moment cannot both claim the same earnings. The ledger
        posting happens here and not when somebody presses "paid" — the platform stops
        owing the moment it decides to pay, and the bank transfer that follows is evidence
        recorded against something that already exists. Nothing owing answers 204: not a
        failure, and not a payout either.
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PreparePayoutCommand' }
      responses:
        '201':
          description: The payout run, awaiting a bank transfer somebody makes by hand.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PartnerPayoutEnvelope' }
        '204': { description: Nothing is owing to that partner. }
        '400': { description: The command is malformed. }
        '401': { description: A valid staff session is required. }
        '403': { description: The account may not settle partner payouts. }
        '503': { description: The payout desk is temporarily unavailable. }
  /api/v1/admin/settlement/payouts/{payoutId}/paid:
    post:
      operationId: markPartnerPayoutPaid
      summary: Record that the money was actually sent.
      description: >-
        Nothing here moves money. The platform holds no banking credentials; a person
        makes the transfer and comes back to record what the bank called it, which is the
        only thing that lets the payment be found again on a statement. Recording the same
        payout twice is the same outcome, not a second transfer.
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters:
        - name: payoutId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BankReferenceCommand' }
      responses:
        '200':
          description: The payout, now settled.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PartnerPayoutEnvelope' }
        '400': { description: The bank reference is missing or too long. }
        '401': { description: A valid staff session is required. }
        '403': { description: The account may not settle partner payouts. }
        '404': { description: No such payout in this tenant. }
        '409': { description: That payout is no longer awaiting payment. }
        '503': { description: The payout desk is temporarily unavailable. }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: alo_session
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  schemas:
    ResponseMeta:
      type: object
      required: [requestId, timestamp, version]
      properties:
        requestId: { type: string, format: uuid }
        timestamp: { type: string, format: date-time }
        version: { type: string, const: v1 }
...
    Money:
      type: object
      required: [amount, currency]
      properties:
        amount: { type: string, pattern: '^\d+$' }
        currency: { type: string, enum: [IRR] }
...
    AdminRoleSummary:
      type: object
      required: [code, name, permissions, grantable]
      properties:
        code: { type: string, maxLength: 64 }
        name: { type: string, maxLength: 120 }
        permissions: { type: array, items: { type: string, maxLength: 100 } }
        grantable:
          type: boolean
          description: False when the role carries a permission the caller does not hold.
    StaffMember:
      type: object
      required: [accountId, mobileE164, status, roles, permissions, isSelf]
      properties:
        accountId: { type: string, format: uuid }
        mobileE164: { type: string, maxLength: 16 }
        status: { type: string, enum: [ACTIVE, SUSPENDED, CLOSED] }
        roles:
          type: array
          items:
            type: object
            required: [grantId, code, name, grantedAt, expiresAt]
            properties:
              grantId: { type: string, format: uuid }
              code: { type: string, maxLength: 64 }
              name: { type: string, maxLength: 120 }
              grantedAt: { type: string, format: date-time }
              expiresAt: { type: [string, 'null'], format: date-time }
        permissions:
          type: array
          items: { type: string, maxLength: 100 }
          description: Union of every permission this account's live grants carry.
        isSelf:
          type: boolean
          description: True for the account making the request, which may not revoke itself.
    GrantRoleCommand:
      type: object
      additionalProperties: false
      required: [mobileE164, roleCode, reason]
      properties:
        mobileE164: { type: string, pattern: '^\+989\d{9}$' }
        roleCode: { type: string, pattern: '^[A-Z][A-Z0-9_]{2,63}$' }
        reason: { type: string, minLength: 3, maxLength: 280 }
    RevokeRoleCommand:
      type: object
      additionalProperties: false
      required: [grantId, reason]
      properties:
        grantId: { type: string, format: uuid }
        reason: { type: string, minLength: 3, maxLength: 280 }
    AdminRoleListEnvelope:
      type: object
      required: [success, data, meta]
      properties:
        success: { type: boolean, const: true }
        data:
          type: array
          items: { $ref: '#/components/schemas/AdminRoleSummary' }
        meta: { $ref: '#/components/schemas/ResponseMeta' }
    StaffMemberEnvelope:
      type: object
      required: [success, data, meta]
      properties:
        success: { type: boolean, const: true }
        data: { $ref: '#/components/schemas/StaffMember' }
        meta: { $ref: '#/components/schemas/ResponseMeta' }
    StaffMemberListEnvelope:
      type: object
      required: [success, data, meta]
      properties:
        success: { type: boolean, const: true }
        data:
          type: array
          items: { $ref: '#/components/schemas/StaffMember' }
        meta: { $ref: '#/components/schemas/ResponseMeta' }
...
    GrantableBranch:
      type: object
      required: [id, nameFa, bakeryNameFa]
      properties:
        id: { type: string, format: uuid }
        nameFa: { type: string, minLength: 1 }
        bakeryNameFa: { type: string, minLength: 1 }
    GrantableBranchListEnvelope:
      type: object
      required: [success, data, meta]
      properties:
        success: { type: boolean, const: true }
        data:
          type: array
          items: { $ref: '#/components/schemas/GrantableBranch' }
        meta: { $ref: '#/components/schemas/ResponseMeta' }
    SettlementParty:
      type: string
      enum: [BAKERY, COURIER]
    PartnerBalance:
      type: object
      required: [party, partnerId, partnerName, amount, orderCount, oldestAt]
      properties:
        party: { $ref: '#/components/schemas/SettlementParty' }
        partnerId: { type: string, format: uuid }
        partnerName: { type: string, minLength: 1 }
        amount: { $ref: '#/components/schemas/Money' }
        orderCount: { type: integer, minimum: 0 }
        oldestAt:
          type: string
          format: date-time
          nullable: true
          description: >-
            When the oldest unpaid order was delivered. The figure a total alone hides:
            two partners owed the same amount are in very different positions if one has
            been waiting since last month.
    PartnerBalanceListEnvelope:
      type: object
      required: [success, data, meta]
      properties:
        success: { type: boolean, const: true }
        data:
          type: array
          items: { $ref: '#/components/schemas/PartnerBalance' }
        meta: { $ref: '#/components/schemas/ResponseMeta' }
    PartnerPayoutSummary:
      type: object
      required:
        [
          id,
          party,
          partnerId,
          partnerName,
          state,
          amount,
          orderCount,
          periodStart,
          periodEnd,
          createdAt,
        ]
      properties:
        id: { type: string, format: uuid }
        party: { $ref: '#/components/schemas/SettlementParty' }
        partnerId: { type: string, format: uuid }
        partnerName: { type: string, minLength: 1 }
        state: { type: string, enum: [DRAFT, PAID, CANCELLED] }
        amount: { $ref: '#/components/schemas/Money' }
        orderCount: { type: integer, minimum: 1 }
        periodStart: { type: string, format: date-time }
        periodEnd: { type: string, format: date-time }
        bankReference:
          type: string
          minLength: 1
          maxLength: 128
          description: What the bank called the transfer. Present once somebody sent the money.
        createdAt: { type: string, format: date-time }
        paidAt: { type: string, format: date-time }
    PartnerPayoutEnvelope:
      type: object
      required: [success, data, meta]
      properties:
        success: { type: boolean, const: true }
        data: { $ref: '#/components/schemas/PartnerPayoutSummary' }
        meta: { $ref: '#/components/schemas/ResponseMeta' }
    PartnerPayoutListEnvelope:
      type: object
      required: [success, data, meta]
      properties:
        success: { type: boolean, const: true }
        data:
          type: array
          items: { $ref: '#/components/schemas/PartnerPayoutSummary' }
        meta: { $ref: '#/components/schemas/ResponseMeta' }
    PreparePayoutCommand:
      type: object
      additionalProperties: false
      required: [party, partnerId, idempotencyKey]
      description: >-
        No amount. It is whatever the partner has earned and not been paid at the instant
        the transaction reads it; a staff-supplied figure would be a staff-chosen payout,
        and the one thing a payout must not be is a number somebody typed.
      properties:
        party: { $ref: '#/components/schemas/SettlementParty' }
        partnerId: { type: string, format: uuid }
        idempotencyKey: { type: string, minLength: 16, maxLength: 128 }
    BankReferenceCommand:
      type: object
      additionalProperties: false
      required: [bankReference]
      properties:
        bankReference:
          type: string
          minLength: 1
          maxLength: 128
          description: >-
            The bank's own reference. Required, because a transfer recorded with nothing
            to look it up by cannot be reconciled against a statement, and reconciling
            against a statement is the entire point of recording it.
...